func, callback: instantiate Wasm functions on boot
Previous dynamic lookup of Wasm function was lazy and only performed
on its first use - this is redundant, and the logic is much clearer
when the functions are initialized on startup + when they're registered
dynamically.
