+++ src/callback.c
@@ -16,12 +16,6 @@
 
 #include "sqliteInt.h"
 
-#ifdef LIBSQL_ENABLE_WASM_RUNTIME
-#include <wasm.h>
-#include <wasmtime.h>
-#include <wasmtime/error.h>
-#endif
-
 /*
 ** Invoke the 'collation needed' callback to request a collation sequence
 ** in the encoding enc of name zName, length nName.
@@ -380,148 +374,6 @@ void sqlite3InsertBuiltinFuncs(
 
 
 
-#ifdef LIBSQL_ENABLE_WASM_RUNTIME
-void run_wasm(sqlite3_context *context, int argc, sqlite3_value **argv);
-
-static void free_wasm_module(void *module) {
-  wasmtime_module_delete(*(wasmtime_module_t **)module);
-}
-
-// Tries to initialize and register a Wasm function dynamically.
-// Returns NULL on failure and fills err_msg_buf with an error message, if not NULL.
-FuncDef *try_instantiate_wasm_function(sqlite3 *db, const char *pName, int nName, const char *pSrcBody, int nBody, int nArg, wasm_byte_vec_t *err_msg_buf) {
-  FuncDef *pBest = NULL;
-  FuncDef *pOther = NULL;
-
-  if (!db->wasm.engine) {
-    db->wasm.engine = wasm_engine_new();
-  }
-  wasm_byte_vec_t compiled_wasm;
-  wasmtime_error_t *error = wasmtime_wat2wasm(pSrcBody, nBody, &compiled_wasm);
-  if (error) {
-    if (err_msg_buf) {
-      wasmtime_error_message(error, err_msg_buf);
-    }
-    return NULL;
-  }
-
-  wasmtime_module_t *module = NULL;
-  error = wasmtime_module_new(db->wasm.engine, (uint8_t *)compiled_wasm.data, compiled_wasm.size, &module);
-  wasm_byte_vec_delete(&compiled_wasm);
-  if (error) {
-    if (err_msg_buf) {
-      wasmtime_error_message(error, err_msg_buf);
-    }
-    return NULL;
-  }
-
-  pBest = sqlite3DbMallocZero(db, sizeof(*pBest)+sizeof(module)+nName+1);
-  memcpy(&pBest[1], &module, sizeof(module));
-  pBest->zName = (const char*)&pBest[1] + sizeof(module);
-  memcpy((char *)pBest->zName, pName, nName);
-  ((char *)pBest->zName)[nName] = '\0';
-
-  pBest->funcFlags = 0;
-  pBest->xSFunc = run_wasm;
-  pBest->xFinalize = NULL;
-  pBest->xValue = NULL;
-  pBest->xInverse = NULL;
-  pBest->pUserData = &pBest[1];
-  pBest->nArg = (i8)nArg;
-
-  pBest->u.pDestructor = (FuncDestructor *)sqlite3Malloc(sizeof(FuncDestructor));
-  if(!pBest->u.pDestructor) {
-    sqlite3OomFault(db);
-    free_wasm_module(pBest);
-    return NULL;
-  }
-  pBest->u.pDestructor->nRef = 1;
-  pBest->u.pDestructor->xDestroy = free_wasm_module;
-  pBest->u.pDestructor->pUserData = &pBest[1];
-
-  // Register the function dynamically
-  for(u8 *z = (u8*)pBest->zName; *z; z++) {
-    *z = sqlite3UpperToLower[*z];
-  };
-  pOther = (FuncDef*)sqlite3HashInsert(&db->aFunc, pBest->zName, pBest);
-  if (pOther == pBest) {
-    sqlite3DbFree(db, pBest);
-    sqlite3OomFault(db);
-    return NULL;
-  } else {
-    pBest->pNext = pOther;
-  }
-
-  return pBest;
-}
-
-void functionDestroy(sqlite3 *db, FuncDef *p);
-
-// Removes a Wasm function
-int deregister_wasm_function(sqlite3 *db, const char *zName) {
-  FuncDef *p = (FuncDef *)sqlite3HashInsert(&db->aFunc, zName, NULL);
-  int ret = p ? 1 : 0;
-  while (p) {
-    functionDestroy(db, p);
-    FuncDef *pNext = p->pNext;
-    sqlite3DbFree(db, p);
-    p = pNext;
-  };
-  return ret;
-}
-
-static FuncDef *maybe_instantiate_wasm_function(sqlite3 *db, const char *zName, int nName, int nArg) {
-  FuncDef *pBest = NULL;
-  const unsigned char *pSrcBody = NULL;
-  int nSrcBody = 0;
-  sqlite3_stmt *stmt;
-  int rc;
-
-  if (!db->aDb || !db->aDb->pSchema) {
-    return NULL;
-  }
-  Table *tab = sqlite3HashFind(&db->aDb->pSchema->tblHash, "libsql_wasm_func_table");
-  if (!tab) {
-    return NULL;
-  }
-  rc = sqlite3_prepare_v2(db, "SELECT body FROM libsql_wasm_func_table WHERE name = ?", -1, &stmt, 0);
-  if (rc != SQLITE_OK) {
-    sqlite3_finalize(stmt);
-    return NULL;
-  }
-  rc = sqlite3_bind_text(stmt, 1, zName, -1, SQLITE_STATIC);
-  if (rc != SQLITE_OK) {
-    sqlite3_finalize(stmt);
-    return NULL;
-  }
-
-  rc = sqlite3_step(stmt);
-  if (rc != SQLITE_ROW) {
-    sqlite3_finalize(stmt);
-    return NULL;
-  }
-
-  pSrcBody = sqlite3_column_text(stmt, 0);
-  nSrcBody = sqlite3_column_bytes(stmt, 0);
-
-  wasm_name_t *err_msg_buf = NULL;
-#ifdef SQLITE_DEBUG
-  wasm_byte_vec_t err_msg;
-  err_msg_buf = &err_msg;
-#endif
-  pBest = try_instantiate_wasm_function(db, zName, nName, pSrcBody, nSrcBody, nArg, err_msg_buf);
-#ifdef SQLITE_DEBUG
-  if (!pBest) {
-    sqlite3DebugPrintf("Failed to instantiate Wasm function: %s\n", err_msg.data);
-    wasm_byte_vec_delete(err_msg_buf);
-  }
-#endif
-  sqlite3_finalize(stmt);
-
-  return pBest;
-}
-#endif // LIBSQL_ENABLE_WASM_RUNTIME
-
 /*
 ** Locate a user function given a name, a number of arguments and a flag
 ** indicating whether the function prefers UTF-16 over UTF-8.  Return a
@@ -622,11 +474,6 @@ FuncDef *sqlite3FindFunction(
   if( pBest && (pBest->xSFunc || createFlag) ){
     return pBest;
   }
-
-#ifdef LIBSQL_ENABLE_WASM_RUNTIME
-  return maybe_instantiate_wasm_function(db, zName, nName, nArg);
-#endif
-
   return 0;
 }
