+++ src/func.c
@@ -2403,6 +2403,116 @@ void run_wasm(sqlite3_context *context, int argc, sqlite3_value **argv) {
     sqlite3_result_error(context, "Wasm function returned unsupported result type", -1);
   }
 }
+
+static void free_wasm_module(void *module) {
+  wasmtime_module_delete(*(wasmtime_module_t **)module);
+}
+
+// Tries to initialize and register a Wasm function dynamically.
+// Returns NULL on failure and fills err_msg_buf with an error message, if not NULL.
+FuncDef *try_instantiate_wasm_function(sqlite3 *db, const char *pName, int nName, const char *pSrcBody, int nBody, int nArg, wasm_byte_vec_t *err_msg_buf) {
+  FuncDef *pBest = NULL;
+  FuncDef *pOther = NULL;
+
+  if (!db->wasm.engine) {
+    db->wasm.engine = wasm_engine_new();
+  }
+  wasm_byte_vec_t compiled_wasm;
+  wasmtime_error_t *error = wasmtime_wat2wasm(pSrcBody, nBody, &compiled_wasm);
+  if (error) {
+    if (err_msg_buf) {
+      wasmtime_error_message(error, err_msg_buf);
+    }
+    return NULL;
+  }
+
+  wasmtime_module_t *module = NULL;
+  error = wasmtime_module_new(db->wasm.engine, (uint8_t *)compiled_wasm.data, compiled_wasm.size, &module);
+  wasm_byte_vec_delete(&compiled_wasm);
+  if (error) {
+    if (err_msg_buf) {
+      wasmtime_error_message(error, err_msg_buf);
+    }
+    return NULL;
+  }
+
+  pBest = sqlite3DbMallocZero(db, sizeof(*pBest)+sizeof(module)+nName+1);
+  memcpy(&pBest[1], &module, sizeof(module));
+  pBest->zName = (const char*)&pBest[1] + sizeof(module);
+  memcpy((char *)pBest->zName, pName, nName);
+  ((char *)pBest->zName)[nName] = '\0';
+
+  pBest->funcFlags = 0;
+  pBest->xSFunc = run_wasm;
+  pBest->xFinalize = NULL;
+  pBest->xValue = NULL;
+  pBest->xInverse = NULL;
+  pBest->pUserData = &pBest[1];
+  pBest->nArg = (i8)nArg;
+
+  pBest->u.pDestructor = (FuncDestructor *)sqlite3Malloc(sizeof(FuncDestructor));
+  if(!pBest->u.pDestructor) {
+    sqlite3OomFault(db);
+    free_wasm_module(pBest);
+    return NULL;
+  }
+  pBest->u.pDestructor->nRef = 1;
+  pBest->u.pDestructor->xDestroy = free_wasm_module;
+  pBest->u.pDestructor->pUserData = &pBest[1];
+
+  // Register the function dynamically
+  for(u8 *z = (u8*)pBest->zName; *z; z++) {
+    *z = sqlite3UpperToLower[*z];
+  };
+  pOther = (FuncDef*)sqlite3HashInsert(&db->aFunc, pBest->zName, pBest);
+  if (pOther == pBest) {
+    sqlite3DbFree(db, pBest);
+    sqlite3OomFault(db);
+    return NULL;
+  } else {
+    pBest->pNext = pOther;
+  }
+
+  return pBest;
+}
+
+void functionDestroy(sqlite3 *db, FuncDef *p);
+
+// Removes a Wasm function
+int deregister_wasm_function(sqlite3 *db, const char *zName) {
+  FuncDef *p = (FuncDef *)sqlite3HashInsert(&db->aFunc, zName, NULL);
+  int ret = p ? 1 : 0;
+  while (p) {
+    functionDestroy(db, p);
+    FuncDef *pNext = p->pNext;
+    sqlite3DbFree(db, p);
+    p = pNext;
+  };
+  return ret;
+}
+
+// This initialization is best-effort: failing to initialize a Wasm func table
+// does not prevent users from using this database
+
+FuncDef *try_instantiate_wasm_function(sqlite3 *db, const char *pName, int nName, const char *pSrcBody, int nBody, int nArg, wasm_byte_vec_t *err_msg_buf);
+int instantiate_callback(void *db, int argc, char **argv, char **names) {
+  try_instantiate_wasm_function((sqlite3 *)db, argv[0], sqlite3Strlen30(argv[0]), argv[1], sqlite3Strlen30(argv[1]), -1, NULL);
+  return SQLITE_OK;
+}
+
+void libsql_try_initialize_wasm_func_table(sqlite3 *db) {
+#ifdef SQLITE_TEST
+  // Tcl tests assume there are no extra tables during their execution
+  return;
+#endif
+  int rc = sqlite3_exec(db, "CREATE TABLE IF NOT EXISTS libsql_wasm_func_table (name text PRIMARY KEY, body text)", NULL, NULL, NULL);
+  if (rc == SQLITE_OK) {
+    // If the table exists, register all the functions
+    sqlite3_exec(db, "SELECT name, body FROM libsql_wasm_func_table", instantiate_callback, db, NULL);
+  }
+  sqlite3_exec(db, "SELECT 0", NULL, NULL, NULL);
+}
+
 #endif // LIBSQL_ENABLE_WASM_RUNTIME
 
 /*
@@ -2590,16 +2700,3 @@ void sqlite3RegisterBuiltinFunctions(void){
   }
 #endif
 }
-
-#ifdef LIBSQL_ENABLE_WASM_RUNTIME
-int libsql_initialize_wasm_func_table(sqlite3 *db) {
-  int rc = sqlite3_exec(db, "CREATE TABLE IF NOT EXISTS libsql_wasm_func_table (name text PRIMARY KEY, body text)", NULL, NULL, NULL);
-  if (rc == SQLITE_READONLY) {
-    sqlite3_exec(db, "CREATE TEMP TABLE IF NOT EXISTS libsql_wasm_func_table (name text PRIMARY KEY, body text)", NULL, NULL, NULL);
-  }
-  // This initialization is best-effort: failing to initialize a Wasm func table
-  // does not prevent users from using this database
-  db->errCode = SQLITE_OK;
-  return SQLITE_OK;
-}
-#endif
